requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.3.2
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.40.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imbalanced-learn==0.11.0
imblearn==0.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
pyspark==3.4.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
recursion-visualiser==1.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
statsmodels==0.14.0
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
wsproto==1.2.0
